Extract getParaTimeName (some tests mock useParaTimes.ts file)
