+++ src/app/pages/ParaTimesPage/ParaTimeSelection/index.tsx
@@ -7,7 +7,8 @@ import { Select } from 'grommet/es6/components/Select'
 import { useTranslation } from 'react-i18next'
 import { ParaTimeContent } from '../ParaTimeContent'
 import { ParaTimeFormFooter } from '../ParaTimeFormFooter'
-import { getParaTimeName, useParaTimes } from '../useParaTimes'
+import { useParaTimes } from '../useParaTimes'
+import { getParaTimeName } from '../getParaTimeName'
 import { useParaTimesNavigation } from '../useParaTimesNavigation'
 import { ThemeContext } from 'styled-components'
 import { ThemeType } from 'grommet/es6/themes'
